Refactors contact page for server actions
